requirements-dev.txt
alembic==1.5.7
apispec==4.4.0
attrs==20.3.0
autopep8==1.5.6
blinker==1.4
Bootstrap-Flask==1.5.1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cryptography==3.4.7
currency-symbols==2.0.0
dnspython==2.1.0
email-validator==1.1.2
flake8==3.9.0
Flask==1.1.2
Flask-Admin==1.5.7
Flask-Dance==3.3.1
Flask-DebugToolbar==0.11.0
Flask-JWT-Extended==4.1.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-Moment==0.11.0
flask-smorest==0.30.0
Flask-SQLAlchemy==2.5.1
flask-whooshalchemy3==0.2.0
Flask-WTF==0.14.3
greenlet==1.0.0
idna==2.7
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.11.1
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.9
pbr==3.1.1
pep517==0.10.0
pip-tools==6.1.0
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.0
PyJWT==2.0.1
pyngrok==5.0.5
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
PyYAML==5.4.1
requests==2.20.0
requests-oauthlib==1.3.0
restcountries==1.1.3
six==1.11.0
SQLAlchemy==1.3.24
toml==0.10.2
urllib3==1.24.3
URLObject==2.4.3
webargs==8.0.0
Werkzeug==1.0.1
Whoosh==2.7.4
WTForms==2.3.3
sentry-sdk==1.0.0
